+++ src/number/util/coerce.js
@@ -1,3 +1,0 @@
-export default function(n) {
-  return n != null ? +n : n;
-}